Update compile settings for paragraph tabs
Adding a tab on the first paragraph now has separate options for
the first paragraph of the story, per chapter, and per section. There
is a separate new compile.paragraph.mode option that determines if
paragraphs are justified (no tab on the first line with blank lines
separating paragraphs) vs prose-style (a tab on the first line per
paragraph with no blank lines between paragraphs).